requirements.txt
autopep8==1.6.0
bcrypt==3.2.0
blinker==1.4
cffi==1.15.0
click==8.0.4
dnspython==2.2.0
email-validator==1.1.3
Flask==2.0.3
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.0.1
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-http-client==3.3.6
six==1.16.0
SQLAlchemy==1.4.31
starkbank-ecdsa==2.0.3
toml==0.10.2
Werkzeug==2.0.3
WTForms==3.0.1
